Moving flag as deleted behavior to a redis set.
This is a large refactor combined with new functionality.
I've removed the flag as deleted job and rolled that functionality into
the current delta job.  When a model is indexed I create the delta job
and add the flag as deleted document id to a redis set.  The delta job
handles getting the set of document ids to flag as deleted without
racing.

This should also allow multiple resque workers to work off different
delta index jobs (on distinct indices) concurrently.