Fix feature:
I found that this scenario passed regardless of whether or not the line
`And I run the smart indexer` was present. I figured we might as well
test to see whether if starting the smart indexer before the initial
delta job gets a chance to run ends up with the correct results.